+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added Vim mode
 - Added theme toggle
 - Fixed titlebar icon color
+- Added togglable status bar
+- Added optional word count in status bar
 
 ### Released